Try at fixing 'cast to/from pointer to/from integer of different size' warnings
git-svn-id: svn://svn.rockbox.org/rockbox/trunk@21719 a1c6a512-1295-4272-9138-f99709370657

+++ apps/codecs/cook.c
@@ -34,7 +34,7 @@ COOKContext q;
 
 static void init_rm(RMContext *rmctx)
 {
-    memcpy(rmctx, (void*)(( (int)ci->id3->id3v2buf + 3 ) &~ 3), sizeof(RMContext));
+    memcpy(rmctx, (void*)(( (intptr_t)ci->id3->id3v2buf + 3 ) &~ 3), sizeof(RMContext));
 }
 
 /* this is the codec entry point */
